Refactor extensions and tests
